% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODWGviz.r
\docType{package}
\name{ODWGviz-package}
\alias{ODWGviz}
\alias{ODWGviz-package}
\title{ODWGviz}
\description{
Package for testing and visualizing outlier detection algorithms.
}
\author{
\strong{Maintainer}: Michael Koohafkan \email{michael.koohafkan@water.ca.gov}

}
